# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart LPR API",
        "description": "Open API for Smart LPR",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/history/{date_start}/{date_end}/{gate}/{page}/{limit}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get history of items",
                "description": "Retrieve a history of items based on date range, gate, pagination, and limit",
                "operationId": "history",
                "parameters": [
                    {
                        "name": "date_start",
                        "in": "path",
                        "description": "Start date for filtering the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "datetime",
                            "example": "2024-09-01 00:00:00"
                        }
                    },
                    {
                        "name": "date_end",
                        "in": "path",
                        "description": "End date for filtering the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "datetime",
                            "example": "2024-09-05 23:59:59"
                        }
                    },
                    {
                        "name": "gate",
                        "in": "path",
                        "description": "Gate number for filtering the list",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "66d9965ec312805ebe04d8bd"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/history": {
            "post": {
                "tags": [
                    "History"
                ],
                "summary": "Store a newly created history in storage",
                "operationId": "d276b0f967aaec24c676e1290f07e897",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "lpr": {
                                        "description": "License plate recognition (LPR) value",
                                        "type": "string",
                                        "example": "ABC1234"
                                    },
                                    "type": {
                                        "description": "Type of the list entry",
                                        "type": "string",
                                        "example": "Vehicle"
                                    },
                                    "traffic_type": {
                                        "description": "Type of traffic (e.g., 'Entry' or 'Exit')",
                                        "type": "string",
                                        "example": "Entry"
                                    },
                                    "image": {
                                        "description": "Image file associated with the entry",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "gate_id": {
                                        "description": "ID of the gate where the traffic occurred",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "timestamp": {
                                        "description": "Timestamp when the entry was recorded",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-09-04T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "History created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/history/{id}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Display the specified history",
                "operationId": "c9c548a03de9b70d773d28eb6c0dd6f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the history to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "History retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}